CalculateAxisPath {ggradar}	R Documentation

Calculate Axis Path

Description

Calculates x-y coordinates for a set of radial axes (one per variable being plotted in radar plot)

Usage

CalculateAxisPath(var.names, min, max)

Arguments

var.names
list of variables to be plotted on radar plot

min
MININUM value required for the plotted axes (same value will be applied to all axes)

max
MAXIMUM value required for the plotted axes (same value will be applied to all axes)

Value

a dataframe of the calculated axis paths

[Package ggradar version 0.2 Index]