CalculateGroupPath {ggradar}	R Documentation

Calculate Group Path

Description

Converts variable values into a set of radial x-y coordinates

Usage

CalculateGroupPath(df)

Arguments

df
a dataframe with Col 1 is group ('unique' cluster / group ID of entity) and Col 2-n are v1.value to vn.value - values (e.g. group/cluser mean or median) of variables v1 to v.n

Value

a dataframe of the calculated axis paths

Source

Code adapted from a solution posted by Tony M to http://stackoverflow.com/questions/9614433/creating-radar-chart-a-k-a-star-plot-spider-plot-using-ggplot2-in-r.

[Package ggradar version 0.2 Index]